Move DelayedRender component to its own module
